+++ windows/src/cascadia/TerminalSettingsEditor/pch.h
@@ -43,6 +43,7 @@
 #include <winrt/Windows.UI.Xaml.Markup.h>
 #include <winrt/Windows.UI.Xaml.Media.h>
 #include <winrt/Windows.UI.Xaml.Navigation.h>
+#include <winrt/Windows.UI.Xaml.Shapes.h>
 
 #include <winrt/Windows.Media.h>
 #include <winrt/Windows.Media.Core.h>